feat: Added support for multiple preferred languages
- SupportedLanguage.fromDeviceLanguages() now returns an array of
SupportedLanguage, containing all the SupportedLanguages that also
appear inside the User's preferred languages list, preserving the
order.
- SupportedLanguage.fromDeviceLanguage() now returns the first
language that matches among those 2 lists: if the first preferred
language is not supported but the second one is, it will now return
that one.
It still returns .en if there is no match between the two lists.